pyproject.toml
[tool.poetry]
name = "read-agent-java"
version = "0.1.0"
description = "AI assistant to help grooming Java project development task"
authors = ["Jerry You <jy2947@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
anthropic = "0.31.0"
jira = "3.8.0"
langfuse = "2.40.0"
pytest = "8.0.2"
openai = "^1.40.3"
google-cloud-aiplatform = "^1.62.0"
google-generativeai = "^0.7.2"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
